fix: integer overflow

+++ libs/processing/align_images.py
@@ -60,7 +60,7 @@ def find_corners(image, filter_grayscale, num=10, gray_filter=20):
     for contour in contours:
         rect = cv2.minAreaRect(contour)
         bounding_box = cv2.boxPoints(rect)
-        box_points += list(np.int8(bounding_box))
+        box_points += list(np.int32(bounding_box))
 
     copy_image = image.copy()
     for corner in box_points:
